requirements.txt
pandas==2.2.0
requests==2.32.0
beautifulsoup4==4.12.0
google-auth==2.36.0
google-api-python-client==2.152.0
sqlalchemy==2.0.0
psycopg2-binary==2.9.0
pytest-cov==6.0.0
pytest==8.0.0
coverage==7.0.0